WIP: Addressed PR comments.
