+++ buffer.c
@@ -1,4 +1,4 @@
-#include "mkmimo.h"
+#include "utils.h"
 
 Buffer *new_buffer() {
   Buffer *buf = malloc(sizeof(Buffer));
