+++ mkmimo.h
@@ -25,14 +25,6 @@
 #define DEFAULT_BLOCKSIZE (4 * BUFSIZ) // 4096
 extern int BLOCKSIZE;
 
-struct input;
-typedef struct input_buffer {
-  void *data;
-  int capacity;
-  int begin, size;        // Byte ranging containing data
-  int end_of_last_record; // Last record seperator found in range
-} Buffer;
-
 typedef struct input {
   int fd;
   char *name;
