--- utils.h~
@@ -1,2 +1,0 @@
-Buffer *new_buffer();
-void enlarge_buffer(Buffer *buf, size_t new_capacity);
